--- heroku.yml
@@ -1,3 +1,0 @@
-build:
-  docker:
-    web: Dockerfile
