--- text-generator-py/.flaskenv
@@ -1,4 +1,0 @@
-FLASK_APP=text_generator
-FLASK_ENV=development
-FLASK_DEBUG=0
-PYTHONUNBUFFERED=0
