--- text-generator-py/docker-compose.yml
@@ -1,9 +1,0 @@
-version: "3.8"
-services:
-  app:
-    image: text_generator
-#    command: flask run
-    ports:
-      - "5000:5000"
-    env_file:
-      - .flaskenv
