--- text-generator-py/requirements.txt
@@ -1,6 +1,0 @@
-flask
-fire>=0.1.3
-regex==2017.4.5
-requests==2.21.0
-tqdm==4.31.1
-toposort==1.5
