--- text-generator-py/text_generator/.gitignore
@@ -1,5 +1,0 @@
-__pycache__
-.mypy_cache/
-models/
-checkpoint
-samples
